requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
plotly.express
scipy
statsmodels